requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.3.0
gunicorn==21.2.0
